+++ docs/changelog.md
@@ -22,6 +22,7 @@
 - Fixed #120 - Image files should now be found under Linux
 - Fixed #128 - Sometimes the false platform image where deleted
 - Fixed #129 - NOTFORSALE and ONLYMOBILE parameters are now read correctly when publishers are imported
+- Fixed #132 - Genre import: Good/bad gameplay features, and good themes where not imported correctly
 - Hardware import/export: ONLY_STATIONARY and ONLY_HANDHELD parameters where not properly read
 - Hardware Feature: need internet was not applied when hardware feature was added
 - Hardware Feature import/export: NEEDINTERNET parameter was not properly read